# src/infrastructure/express/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ProjectResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "API",
		"version": "1.0.0"
	},
	"openapi": "3.0.0",
	"paths": {
		"/internal/projects": {
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"200": {
						"description": "Return a project",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponse"
								}
							}
						}
					}
				},
				"tags": [
					"project-internal"
				],
				"security": [
					{
						"AUTH0_USER": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProjectCreateParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetProjects",
				"responses": {
					"200": {
						"description": "Return projects",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProjectResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"project-internal"
				],
				"security": [
					{
						"AUTH0_USER": []
					}
				],
				"parameters": []
			}
		},
		"/internal/projects/{id}": {
			"get": {
				"operationId": "FindProject",
				"responses": {
					"200": {
						"description": "Return a project",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponse"
								}
							}
						}
					}
				},
				"tags": [
					"project-internal"
				],
				"security": [
					{
						"AUTH0_USER": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/public/projects/{id}": {
			"get": {
				"operationId": "FindProjectForPublic",
				"responses": {
					"200": {
						"description": "Return a project",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProjectResponse"
								}
							}
						}
					}
				},
				"tags": [
					"project-public"
				],
				"security": [
					{
						"API_KEY": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/public/projects": {
			"get": {
				"operationId": "GetProjectsForPublic",
				"responses": {
					"200": {
						"description": "Return projects",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProjectResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"project-public"
				],
				"security": [
					{
						"API_KEY": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:4000/"
		}
	]
}